+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ## Other changes
 
  * Added support for recording breakdown metrics (#535)
+ * Added support for central config management (#511)
  * Added instrumentation for `urllib2` (Python 2) / `urllib.request` (Python 3) (#464)
  * Added `disable_metrics` setting (#399)
  * Updated elasticsearch instrumentation for 7.x (#482, #483)